First version of 3D reconstruction code
Fixed some issues with the Fourier shifts.

* 3D parameter estimation seems fine
* 3D reconstruction w/o filtering seems to work
* Applying OTFs still seems rather broken, have to check

+++ org/fairsim/linalg/BasicVectors.java
@@ -519,8 +519,10 @@ class BasicVector implements VectorFactory {
 		public void at(int z) {
 		    for (int y=0; y<h; y++)
 		    for (int x=0; x<w; x++) {
-			float phaVal = (float)(2*Math.PI*(
-			    ((float)x*kx)/w+((float)y*ky)/h+((float)z*kz)/d));
+			
+			float phaVal = (float)
+			( 2*Math.PI* (x*kx/w + y*ky/h + z*kz/d) );
+			
 			float si,co;
 			if (fast) {
 			    co = (float)MTool.fcos( phaVal );
